+++ lib/src/i18n/strings_ro.dart
@@ -49,4 +49,10 @@ class _StringsRo extends _StringsI18n {
   List<String> getWeeksShort() {
     return null;
   }
+
+  @override
+  List<String> getMonthsShort() {
+    // TODO: implement getMonthsShort
+    return null;
+  }
 }